feat: fix bug

+++ src/utils/LocalStorageHandler.js
@@ -1,17 +1,22 @@
 export const fetchUser = () => {
-  const userInfo =
-    localStorage.getItem("user") !== "undefined"
-      ? JSON.parse(localStorage.getItem("user"))
-      : localStorage.clear();
+  if (typeof window !== "undefined") {
+    const userInfo =
+      localStorage.getItem("user") !== "undefined"
+        ? JSON.parse(localStorage.getItem("user"))
+        : localStorage.clear();
 
-  return userInfo;
+    return userInfo;
+  }
 };
 
 export const fetchCart = () => {
-  const cartInfo =
-    localStorage.getItem("cartItems") !== "undefined"
-      ? JSON.parse(localStorage.getItem("cartItems"))
-      : localStorage.clear();
+  if (typeof window !== "undefined") {
+    // Perform localStorage action
+    const cartInfo =
+      localStorage.getItem("cartItems") !== "undefined"
+        ? JSON.parse(localStorage.getItem("cartItems"))
+        : localStorage.clear();
 
-  return cartInfo ? cartInfo : [];
+    return cartInfo ? cartInfo : [];
+  }
 };
